fix : SentryClient refactoring, move the protocol to its own dart files/classes fix #84
