+++ dart/lib/src/model.dart
@@ -1,0 +1,13 @@
+export 'model/app.dart';
+export 'model/breadcrumb.dart';
+export 'model/browser.dart';
+export 'model/contexts.dart';
+export 'model/device.dart';
+export 'model/dsn.dart';
+export 'model/event.dart';
+export 'model/level.dart';
+export 'model/package.dart';
+export 'model/runtime.dart';
+export 'model/sdk.dart';
+export 'model/system.dart';
+export 'model/user.dart';
