+++ dart/lib/src/model/user.dart
@@ -1,0 +1,57 @@
+/// Describes the current user associated with the application, such as the
+/// currently signed in user.
+///
+/// The user can be specified globally in the [SentryClient.userContext] field,
+/// or per event in the [Event.userContext] field.
+///
+/// You should provide at least either an [id] (a unique identifier for an
+/// authenticated user) or [ipAddress] (their IP address).
+///
+/// Conforms to the User Interface contract for Sentry
+/// https://docs.sentry.io/clientdev/interfaces/user/.
+///
+/// The outgoing JSON representation is:
+///
+/// ```
+/// "user": {
+///   "id": "unique_id",
+///   "username": "my_user",
+///   "email": "foo@example.com",
+///   "ip_address": "127.0.0.1",
+///   "subscription": "basic"
+/// }
+/// ```
+class User {
+  /// At a minimum you must set an [id] or an [ipAddress].
+  const User({this.id, this.username, this.email, this.ipAddress, this.extras})
+      : assert(id != null || ipAddress != null);
+
+  /// A unique identifier of the user.
+  final String id;
+
+  /// The username of the user.
+  final String username;
+
+  /// The email address of the user.
+  final String email;
+
+  /// The IP of the user.
+  final String ipAddress;
+
+  /// Any other user context information that may be helpful.
+  ///
+  /// These keys are stored as extra information but not specifically processed
+  /// by Sentry.
+  final Map<String, dynamic> extras;
+
+  /// Produces a [Map] that can be serialized to JSON.
+  Map<String, dynamic> toJson() {
+    return <String, dynamic>{
+      'id': id,
+      'username': username,
+      'email': email,
+      'ip_address': ipAddress,
+      'extras': extras,
+    };
+  }
+}
