+++ dart/test/sentry_io_test.dart
@@ -2,12 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 @TestOn('vm')
-
 import 'dart:io';
 
 import 'package:sentry/sentry.dart';
+import 'package:sentry/src/io_client.dart';
 import 'package:test/test.dart';
-import 'package:sentry/src/io.dart';
 
 import 'test_utils.dart';
